refactor the code[add render_xml_node]
